requirements.txt
# requirements.txt
Pillow>=9.0.0
hyperlpr3>=0.1.3
